+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - **http client factory:** expose `Count` to know how many clients are registered
 - **http client builder:** auto register client into factory - can be omitted via `WithAutoRegisterFactory` (or `.Build(skipAutoRegister: true)`)
+- **service collection:** configurable defaults on `services.AddFluentlyHttpClient`
 
 ## [4.0.0](https://github.com/sketch7/FluentlyHttpClient/compare/3.9.6...4.0.0) (2024-07-24)